Refractor Provider API to single Pubsub
